chatbot_theme_identifier/backend/requirements.txt
fastapi
uvicorn
python-multipart
python-docx
pytesseract
Pillow
PyMuPDF
chromadb
sentence-transformers
groq
python-dotenv